requirements.txt
streamlit
torch
torchvision
pillow
ipykernel 
numpy 
matplotlib 
seaborn 
pandas